+++ src/Components/Burger/Burger.js
@@ -1,4 +1,6 @@
-import React, { useState } from 'react'
+import React, { useState, useContext } from 'react'
+import BurgerContext from '../../Contex/BurgerContext';
+
 import Bun from './Bun';
 import Tomato from './Tomato';
 import Patty from './Patty';
@@ -6,7 +8,7 @@ import Cheese from './Cheese';
 import Lettuce from './Lettuce';
 import style from './burger.module.css'
 
-const Burger = (props) => {
+const Burger = () => {
 
     // const ingredients = ['patty', 'lettuce', 'bun', 'tomato', 'cheese'];
     // const ingredients = ['patty', 'patty', 'patty', 'patty', 'patty'];
@@ -14,10 +16,12 @@ const Burger = (props) => {
 
     // Urutan burger mengikuti index dari array ingredients diatas
 
+    const ctx = useContext(BurgerContext)
+
     return (
         <div className={style.burger}>
             <Bun type="top"/>   {/* type adalah props dan top adalah valuenya */}
-            {props.ingredients.map( (item,index) => {
+            {ctx.ingredients.map( (item,index) => {
                 switch (item) {
                     case 'patty':
                         return <Patty key={index}/>
