+++ src/Components/Burger/Controller.js
@@ -1,42 +1,47 @@
-import React from 'react'
+import React, { useContext } from 'react'
+import BurgerContext from '../../Contex/BurgerContext'
 
 import AddButton from './AddButton'
 import MapItem from './MapItem'
 import style from './controller.module.css'
 
-const Controller = (props) => {
-  return (
-    <>
-        <div className={style.ingredients_selector}>
-            <h5>Add Ingredients</h5>
-            <div>
-                <AddButton label="Patty" clickHandler={() => props.addIngredientsHandler('patty')} />
-                <AddButton label="Lettuce" clickHandler={() => props.addIngredientsHandler('lettuce')} />
-                <AddButton label="Tomato" clickHandler={() => props.addIngredientsHandler('tomato')} />
-                <AddButton label="Cheese" clickHandler={() => props.addIngredientsHandler('cheese')} />
-                <AddButton label="Bun" clickHandler={() => props.addIngredientsHandler('bun')} />
+const Controller = () => {
+
+    const ctx = useContext(BurgerContext)
+    // console.log('halooo', ctx)
+
+    return (
+        <>
+            <div className={style.ingredients_selector}>
+                <h5>Add Ingredients</h5>
+                <div>
+                    <AddButton label="Patty" clickHandler={() => ctx.addIngredientsHandler('patty')} />
+                    <AddButton label="Lettuce" clickHandler={() => ctx.addIngredientsHandler('lettuce')} />
+                    <AddButton label="Tomato" clickHandler={() => ctx.addIngredientsHandler('tomato')} />
+                    <AddButton label="Cheese" clickHandler={() => ctx.addIngredientsHandler('cheese')} />
+                    <AddButton label="Bun" clickHandler={() => ctx.addIngredientsHandler('bun')} />
+                </div>
             </div>
-        </div>
-        <div className={style.ingredients_map}>
-            {/* Loop untuk membuat list ingredient sesuai dengan urutan array */}
-            {props.ingredients.map( (item, index) => {
-                return (
-                    <>
-                        {/* Merender MapItem untuk setiap item dari 'ingredients' */}
-                        <MapItem
-                            label={item}
-                            key={index}
-                            removeHandler={() => props.removeIngredientsHandler(index)}
-                            upHandler={() => props.moveUpIngredientsHandler(index)}
-                            downHandler={() => props.moveDownIngredientsHandler(index)}
-                            firstItem={ index === 0 }
-                            lastItem={ index === props.ingredients.length -1 }
-                        />
-                    </>
-                )
-            })}
-        </div>
-    </>
+            <div className={style.ingredients_map}>
+                {/* Loop untuk membuat list ingredient sesuai dengan urutan array */}
+                {ctx.ingredients.map( (item, index) => {
+                    return (
+                        <>
+                            {/* Merender MapItem untuk setiap item dari 'ingredients' */}
+                            <MapItem
+                                label={item}
+                                key={index}
+                                removeHandler={() => ctx.removeIngredientsHandler(index)}
+                                upHandler={() => ctx.moveUpIngredientsHandler(index)}
+                                downHandler={() => ctx.moveDownIngredientsHandler(index)}
+                                firstItem={ index === 0 }
+                                lastItem={ index === ctx.ingredients.length -1 }
+                            />
+                        </>
+                    )
+                })}
+            </div>
+        </>
   )
 }
