+++ src/Contex/BurgerContext.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const BurgerContext = React.createContext()
+
+export default BurgerContext
